+++ app/controllers/actions/DataRetrievalAction.scala
@@ -218,8 +218,15 @@ class DataRetrievalImpl(
     }
 }
 
+
+
+
+
+
+
 class RacdacDataRetrievalImpl(
                          @Racdac dataConnector: UserAnswersCacheConnector,
+                         viewConnector: SchemeDetailsReadOnlyCacheConnector,
                          mode: Mode,
                          minimalPsaConnector: MinimalPsaConnector,
                          schemeDetailsConnector: SchemeDetailsConnector,
@@ -228,14 +235,13 @@ class RacdacDataRetrievalImpl(
 
   override protected def transform[A](request: AuthenticatedRequest[A]): Future[OptionalDataRequest[A]] = {
     implicit val hc: HeaderCarrier = HeaderCarrierConverter.fromRequestAndSession(request, request.session)
-    (mode, request.psaId)  match {
-      case (NormalMode | CheckMode, Some(value)) =>
-        getOrCreateOptionalRequest(srn, value.id, refreshData)(request, hc)
-      case (NormalMode | CheckMode, _) => throw MissingPsaIdException
-      case (UpdateMode | CheckUpdateMode, _) =>
+    mode  match {
+      case NormalMode | CheckMode =>
+        createOptionalRequest(dataConnector.fetch(request.externalId), viewOnly = false)(request)
+      case UpdateMode | CheckUpdateMode =>
         (srn, request.psaId) match {
           case (Some(_), Some(psaId)) =>
-            getOrCreateOptionalRequest(srn, psaId.id, refreshData)(request, hc)
+            getOptionalRequest(srn, psaId.id, refreshData)(request, hc)
           case _ => Future(OptionalDataRequest(
             request = request.request,
             externalId = request.externalId,
@@ -246,10 +252,32 @@ class RacdacDataRetrievalImpl(
           ))
         }
     }
-
   }
 
-  private def getOrCreateOptionalRequest[A](srn: Option[String],
+  private def createOptionalRequest[A](f: Future[Option[JsValue]], viewOnly: Boolean)
+    (implicit request: AuthenticatedRequest[A]): Future[OptionalDataRequest[A]] =
+    f.map {
+      case None => OptionalDataRequest(
+        request = request.request,
+        externalId = request.externalId,
+        userAnswers = None,
+        psaId = request.psaId,
+        pspId = request.pspId,
+        viewOnly = viewOnly,
+        administratorOrPractitioner = request.administratorOrPractitioner
+      )
+      case Some(data) => OptionalDataRequest(
+        request = request.request,
+        externalId = request.externalId,
+        userAnswers = Some(UserAnswers(data)),
+        psaId = request.psaId,
+        pspId = request.pspId,
+        viewOnly = viewOnly,
+        administratorOrPractitioner = request.administratorOrPractitioner
+      )
+    }
+
+  private def getOptionalRequest[A](srn: Option[String],
                                        psaId: String,
                                        refresh: Boolean)
                                       (implicit request: AuthenticatedRequest[A],
@@ -276,12 +304,16 @@ class RacdacDataRetrievalImpl(
     }
 
   private def addMinimalFlagsAndUpdateRepository[A](srn: String,
-                                                    jsValue: JsValue,
-                                                    psaId: String)
-                                                   (implicit hc: HeaderCarrier): Future[JsValue] = {
-    minimalPsaConnector.getMinimalFlags(psaId).map { minimalFlags =>
-      UserAnswers(jsValue).set(PsaMinimalFlagsId)(minimalFlags).flatMap(
-        _.set(SchemeSrnId)(srn)).asOpt.getOrElse(UserAnswers(jsValue)).json
+    jsValue: JsValue,
+    psaId: String,
+    upsertUserAnswers: JsValue => Future[JsValue])
+    (implicit hc: HeaderCarrier): Future[JsValue] = {
+    minimalPsaConnector.getMinimalFlags(psaId).flatMap { minimalFlags =>
+      val ua = UserAnswers(jsValue)
+        .set(PsaMinimalFlagsId)(minimalFlags)
+        .flatMap(
+          _.set(SchemeSrnId)(srn)).asOpt.getOrElse(UserAnswers(jsValue))
+      upsertUserAnswers(ua.json)
     }
   }
 
@@ -292,17 +324,16 @@ class RacdacDataRetrievalImpl(
        case (true, Some(value)) =>
        schemeDetailsConnector
          .getSchemeDetails(psaId, schemeIdType = "srn", value)
-         .flatMap(ua => addMinimalFlagsAndUpdateRepository(value, ua.json, psaId))
+         .flatMap(ua => addMinimalFlagsAndUpdateRepository(value, ua.json, psaId, viewConnector.upsert(request.externalId, _)))
          .map(Some(_))
        case _ =>
-         dataConnector.fetch(request.externalId)
+         viewConnector.fetch(request.externalId)
     }
   }
 }
 
 
 case object MissingSchemeNameException extends Exception
-case object MissingPsaIdException extends Exception
 
 @ImplementedBy(classOf[DataRetrievalImpl])
 trait DataRetrieval extends ActionTransformer[AuthenticatedRequest, OptionalDataRequest]
@@ -328,10 +359,11 @@ class DataRetrievalActionImpl @Inject()(dataConnector: UserAnswersCacheConnector
 }
 
 class RacdacDataRetrievalActionImpl @Inject()(@Racdac dataConnector: UserAnswersCacheConnector,
+                                              viewConnector: SchemeDetailsReadOnlyCacheConnector,
                                               minimalPsaConnector: MinimalPsaConnector,
                                               schemeDetailsConnector: SchemeDetailsConnector)(implicit ec: ExecutionContext) extends DataRetrievalAction {
   override def apply(mode: Mode, srn: Option[String], refreshData: Boolean): DataRetrieval = {
-    new RacdacDataRetrievalImpl(dataConnector, mode, minimalPsaConnector, schemeDetailsConnector, srn: Option[String],refreshData: Boolean)
+    new RacdacDataRetrievalImpl(dataConnector, viewConnector, mode, minimalPsaConnector, schemeDetailsConnector, srn: Option[String],refreshData: Boolean)
   }
 }
